Consolidate metered fee types into types.rs
- Move ResourceFeeEstimateResponse and MeteredPriorityFeeResponse to types.rs
- Move build_priority_fee_response helper to meter_rpc.rs
- Delete metered_fee_types.rs
