+++ crates/rpc/src/base/meter_rpc.rs
@@ -18,12 +18,10 @@ use tips_core::types::{Bundle, MeterBundleResponse, ParsedBundle};
 use tracing::{debug, error, info};
 
 use super::{
-    block::meter_block, meter::meter_bundle, traits::MeteringApiServer, types::MeterBlockResponse,
+    block::meter_block, meter::meter_bundle, traits::MeteringApiServer,
+    types::{MeterBlockResponse, MeteredPriorityFeeResponse, ResourceFeeEstimateResponse},
 };
-use crate::{
-    PriorityFeeEstimator, ResourceDemand,
-    base::metered_fee_types::{MeteredPriorityFeeResponse, build_priority_fee_response},
-};
+use crate::{PriorityFeeEstimator, ResourceDemand, ResourceEstimates, RollingPriorityEstimate};
 
 /// Implementation of the metering RPC API
 #[derive(Debug)]
@@ -306,4 +304,35 @@ where
             )
         })
     }
+}
+
+/// Converts a rolling estimate to the response format.
+fn build_priority_fee_response(
+    meter_bundle: MeterBundleResponse,
+    estimate: RollingPriorityEstimate,
+) -> MeteredPriorityFeeResponse {
+    let resource_estimates = build_resource_estimate_responses(&estimate.estimates);
+
+    MeteredPriorityFeeResponse {
+        meter_bundle,
+        priority_fee: estimate.priority_fee,
+        blocks_sampled: estimate.blocks_sampled as u64,
+        resource_estimates,
+    }
+}
+
+fn build_resource_estimate_responses(
+    estimates: &ResourceEstimates,
+) -> Vec<ResourceFeeEstimateResponse> {
+    estimates
+        .iter()
+        .map(|(kind, est)| ResourceFeeEstimateResponse {
+            resource: kind.as_camel_case().to_string(),
+            threshold_priority_fee: est.threshold_priority_fee,
+            recommended_priority_fee: est.recommended_priority_fee,
+            cumulative_usage: U256::from(est.cumulative_usage),
+            threshold_tx_count: est.threshold_tx_count.try_into().unwrap_or(u64::MAX),
+            total_transactions: est.total_transactions.try_into().unwrap_or(u64::MAX),
+        })
+        .collect()
 }
